Add Web Worker for GTM Library.
