+++ .eslintignore
@@ -2,3 +2,5 @@ helix-importer-ui
 blocks/adaptiveform/libs
 blocks/adaptiveform/customization
 blocks/adaptiveform/components
+
+scripts/partytown
