+++ scripts/delayed.js
@@ -24,18 +24,3 @@ loadScript('/scripts/sortable.min.js', {
   type: 'text/javascript',
   charset: 'UTF-8',
 });
-
-//  Load Google Tag Manager
-
-const fireGTM = (w, d, s, l, i) => {
-  w[l] = w[l] || [];
-  w[l].push({ 'gtm.start': new Date().getTime(), event: 'gtm.js' });
-  const f = d.getElementsByTagName(s)[0];
-  const j = d.createElement(s);
-  const dl = l !== 'dataLayer' ? `&l=${l}` : '';
-  j.async = true;
-  j.src = `https://www.googletagmanager.com/gtm.js?id=${i}${dl}`;
-  f.parentNode.insertBefore(j, f);
-};
-
-fireGTM(window, document, 'script', 'dataLayer', 'GTM-WZ7D96C');
