+++ scripts/scripts.js
@@ -15,6 +15,8 @@ import {
   getMetadata,
 } from './lib-franklin.js';
 
+import integrateMartech from './third-party.js';
+
 const LCP_BLOCKS = []; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'airlessco'; // add your RUM generation information here
 
@@ -163,6 +165,17 @@ export function addFavIcon(href) {
 }
 
 /**
+ * Initializes the PartyTown library for processing third-party libraries.
+ */
+function initPartytown() {
+  window.partytown = {
+    lib: '/scripts/partytown/',
+    forward: ['dataLayer.push'],
+  };
+  import('./partytown/partytown.js');
+}
+
+/**
  * loads everything that doesn't need to be delayed.
  */
 async function loadLazy(doc) {
@@ -185,6 +198,9 @@ async function loadLazy(doc) {
   sampleRUM('lazy');
   sampleRUM.observe(main.querySelectorAll('div[data-block-name]'));
   sampleRUM.observe(main.querySelectorAll('picture > img'));
+
+  integrateMartech();
+  initPartytown();
 }
 
 /**
